rename biomass to biomassdata
add comments
